+++ lib/abstracts/json_operations.dart
@@ -1,8 +1,5 @@
 abstract class JsonOperations {
-
   JsonOperations.fromJson(dynamic data);
 
-  toJson()  => {
-
-  };
+  toJson() => {};
 }