+++ lib/metro/models/ny_template.dart
@@ -1,0 +1,13 @@
+/// Used for creating files.
+class NyTemplate {
+  String name;
+  String saveTo;
+  List<String> pluginsRequired = [];
+  String stub;
+
+  NyTemplate(
+      {required this.name,
+      required this.saveTo,
+      required this.pluginsRequired,
+      required this.stub});
+}