+++ lib/router/router.dart
@@ -23,8 +23,8 @@ class NyNavigator {
 }
 
 typedef NyRouteView = Widget Function(
-    BuildContext context,
-    );
+  BuildContext context,
+);
 
 /// Builds the routes in the router.dart file
 NyRouter nyRoutes(Function(NyRouter router) build) {
@@ -34,7 +34,13 @@ NyRouter nyRoutes(Function(NyRouter router) build) {
   return nyRouter;
 }
 
-enum NavigationType { push, pushReplace, pushAndRemoveUntil, popAndPushNamed }
+enum NavigationType {
+  push,
+  pushReplace,
+  pushAndRemoveUntil,
+  popAndPushNamed,
+  pushAndForgetAll
+}
 
 /// NyRouterRoute manages routing, registering routes with transitions, navigating to
 /// routes, closing routes. It is a thin layer on top of [Navigator] to help
@@ -71,14 +77,10 @@ class NyRouter {
   GlobalKey<NavigatorState>? get navigatorKey => _navigatorKey;
 
   /// Get the registered routes names as a list.
-  List<String> getRegisteredRouteNames() {
-    return _routeNameMappings.keys.toList();
-  }
+  List<String> getRegisteredRouteNames() => _routeNameMappings.keys.toList();
 
   /// Get the registered routes as a list.
-  Map<String, NyRouterRoute> getRegisteredRoutes() {
-    return _routeNameMappings;
-  }
+  Map<String, NyRouterRoute> getRegisteredRoutes() => _routeNameMappings;
 
   /// Set the registered routes.
   void setRegisteredRoutes(Map<String, NyRouterRoute> routes) {
@@ -106,8 +108,8 @@ class NyRouter {
   /// Add a new route with a widget.
   route(String name, NyRouteView view,
       {PageTransitionType? transition,
-        PageTransitionSettings? pageTransitionSettings,
-        List<RouteGuard>? routeGuards}) {
+      PageTransitionSettings? pageTransitionSettings,
+      List<RouteGuard>? routeGuards}) {
     this._addRoute(NyRouterRoute(
       name: name,
       view: view,
@@ -120,18 +122,18 @@ class NyRouter {
   /// Add a new page to the router.
   page(NyStatefulWidget widget,
       {PageTransitionType? transition,
-        PageTransitionSettings? pageTransitionSettings,
-        List<RouteGuard>? routeGuards}) {
+      PageTransitionSettings? pageTransitionSettings,
+      List<RouteGuard>? routeGuards}) {
     String widgetRouteName = widget.getRouteName();
     assert(
-    widgetRouteName != "",
-    "${widget.toString()} is missing a route name. \nAdd\n" +
-        """
+        widgetRouteName != "",
+        "${widget.toString()} is missing a route name. \nAdd\n" +
+            """
     @override
     String getRouteName() {
       return "/your-route-name";
     }""" +
-        "\nto your \"${widget.toString()}\" widget class.");
+            "\nto your \"${widget.toString()}\" widget class.");
     this._addRoute(NyRouterRoute(
         name: widgetRouteName,
         view: (context) => widget,
@@ -166,12 +168,12 @@ class NyRouter {
   /// Makes this a callable class. Delegates to [navigate].
   Future<T> call<T>(String name,
       {BaseArguments? args,
-        NavigationType navigationType = NavigationType.push,
-        dynamic result,
-        bool Function(Route<dynamic> route)? removeUntilPredicate,
-        Map<String, dynamic>? params,
-        PageTransitionType? pageTransitionType,
-        PageTransitionSettings? pageTransitionSettings}) async {
+      NavigationType navigationType = NavigationType.push,
+      dynamic result,
+      bool Function(Route<dynamic> route)? removeUntilPredicate,
+      Map<String, dynamic>? params,
+      PageTransitionType? pageTransitionType,
+      PageTransitionSettings? pageTransitionSettings}) async {
     assert(navigationType != NavigationType.pushAndRemoveUntil ||
         removeUntilPredicate != null);
 
@@ -201,18 +203,18 @@ class NyRouter {
   /// [NavigationType.pushAndRemoveUntil] strategy.
   Future<T> navigate<T>(String name,
       {BaseArguments? args,
-        NavigationType navigationType = NavigationType.push,
-        dynamic result,
-        bool Function(Route<dynamic> route)? removeUntilPredicate,
-        PageTransitionType? pageTransitionType,
-        PageTransitionSettings? pageTransitionSettings}) async {
+      NavigationType navigationType = NavigationType.push,
+      dynamic result,
+      bool Function(Route<dynamic> route)? removeUntilPredicate,
+      PageTransitionType? pageTransitionType,
+      PageTransitionSettings? pageTransitionSettings}) async {
     assert(navigationType != NavigationType.pushAndRemoveUntil ||
         removeUntilPredicate != null);
 
     _checkAndThrowRouteNotFound(name, args, navigationType);
 
-    return await _navigate(name, args, navigationType, result, removeUntilPredicate,
-        pageTransitionType, pageTransitionSettings)
+    return await _navigate(name, args, navigationType, result,
+            removeUntilPredicate, pageTransitionType, pageTransitionSettings)
         .then((value) => value as T);
   }
 
@@ -221,8 +223,8 @@ class NyRouter {
   /// [routeArgsPairs] is a list of [RouteArgsPair]. Each [RouteArgsPair]
   /// contains the name of a route and its corresponding argument (if any).
   Future<List> navigateMultiple(
-      List<RouteArgsPair> routeArgsPairs,
-      ) {
+    List<RouteArgsPair> routeArgsPairs,
+  ) {
     assert(routeArgsPairs.isNotEmpty);
 
     final pageResponses = <Future>[];
@@ -321,16 +323,22 @@ class NyRouter {
             .navigatorKey!
             .currentState!
             .popAndPushNamed(name, result: result, arguments: argsWrapper);
+      case NavigationType.pushAndForgetAll:
+        return await this.navigatorKey!.currentState!.pushNamedAndRemoveUntil(
+              name,
+              (_) => false,
+              arguments: argsWrapper,
+            );
     }
   }
 
   /// If the route is not registered throw an error
   /// Make sure to use the correct name while calling navigate.
   void _checkAndThrowRouteNotFound(
-      String name,
-      BaseArguments? args,
-      NavigationType navigationType,
-      ) {
+    String name,
+    BaseArguments? args,
+    NavigationType navigationType,
+  ) {
     if (!_routeNameMappings.containsKey(name)) {
       if (this.options.handleNameNotFoundUI) {
         NyLogger.error("Page not found\n"
@@ -414,23 +422,23 @@ class NyRouter {
       return PageTransition(
           child: StatefulBuilder(
               builder: (BuildContext context, StateSetter setState) {
-                return route.builder(context, baseArgs ?? route.defaultArgs,
-                    queryParameters ?? route.queryParameters);
-              }),
+            return route.builder(context, baseArgs ?? route.defaultArgs,
+                queryParameters ?? route.queryParameters);
+          }),
           type: argumentsWrapper.pageTransitionType ?? route.pageTransitionType,
           settings: settings,
           duration: _getPageTransitionDuration(route, argumentsWrapper),
           alignment: _getPageTransitionAlignment(route, argumentsWrapper),
           reverseDuration:
-          _getPageTransitionReversedDuration(route, argumentsWrapper),
+              _getPageTransitionReversedDuration(route, argumentsWrapper),
           matchingBuilder:
-          _getPageTransitionMatchingBuilder(route, argumentsWrapper),
+              _getPageTransitionMatchingBuilder(route, argumentsWrapper),
           childCurrent: _getPageTransitionChildCurrent(route, argumentsWrapper),
           curve: _getPageTransitionCurve(route, argumentsWrapper),
           ctx: _getPageTransitionContext(route, argumentsWrapper),
           inheritTheme: _getPageTransitionInheritTheme(route, argumentsWrapper),
           fullscreenDialog:
-          _getPageTransitionFullscreenDialog(route, argumentsWrapper),
+              _getPageTransitionFullscreenDialog(route, argumentsWrapper),
           opaque: _getPageTransitionOpaque(route, argumentsWrapper),
           isIos: _getPageTransitionIsIos(route, argumentsWrapper));
     };
@@ -552,7 +560,7 @@ class NyRouter {
   bool _getPageTransitionFullscreenDialog(
       NyRouterRoute route, ArgumentsWrapper argumentsWrapper) {
     bool fullscreenDialog =
-    this.options.pageTransitionSettings.fullscreenDialog!;
+        this.options.pageTransitionSettings.fullscreenDialog!;
 
     if (route.pageTransitionSettings != null &&
         route.pageTransitionSettings!.fullscreenDialog != null) {
@@ -561,7 +569,7 @@ class NyRouter {
     if (argumentsWrapper.pageTransitionSettings != null &&
         argumentsWrapper.pageTransitionSettings!.fullscreenDialog != null) {
       fullscreenDialog =
-      argumentsWrapper.pageTransitionSettings!.fullscreenDialog!;
+          argumentsWrapper.pageTransitionSettings!.fullscreenDialog!;
     }
     return fullscreenDialog;
   }
@@ -602,7 +610,7 @@ class NyRouter {
   PageTransitionsBuilder _getPageTransitionMatchingBuilder(
       NyRouterRoute route, ArgumentsWrapper argumentsWrapper) {
     PageTransitionsBuilder matchingBuilder =
-    this.options.pageTransitionSettings.matchingBuilder!;
+        this.options.pageTransitionSettings.matchingBuilder!;
 
     if (route.pageTransitionSettings != null &&
         route.pageTransitionSettings!.matchingBuilder != null) {
@@ -611,7 +619,7 @@ class NyRouter {
     if (argumentsWrapper.pageTransitionSettings != null &&
         argumentsWrapper.pageTransitionSettings!.matchingBuilder != null) {
       matchingBuilder =
-      argumentsWrapper.pageTransitionSettings!.matchingBuilder!;
+          argumentsWrapper.pageTransitionSettings!.matchingBuilder!;
     }
     return matchingBuilder;
   }
@@ -645,20 +653,20 @@ class NyRouter {
 /// See https://pub.dev/packages/page_transition to learn more.
 routeTo(String routeName,
     {dynamic data,
-      NavigationType navigationType = NavigationType.push,
-      dynamic result,
-      bool Function(Route<dynamic> route)? removeUntilPredicate,
-      PageTransitionSettings? pageTransitionSettings,
-      PageTransitionType? pageTransition,
-      Function(dynamic value)? onPop}) async {
+    NavigationType navigationType = NavigationType.push,
+    dynamic result,
+    bool Function(Route<dynamic> route)? removeUntilPredicate,
+    PageTransitionSettings? pageTransitionSettings,
+    PageTransitionType? pageTransition,
+    Function(dynamic value)? onPop}) async {
   NyArgument nyArgument = NyArgument(data);
   await NyNavigator.instance.router
       .navigate(routeName,
-      args: nyArgument,
-      navigationType: navigationType,
-      result: result,
-      removeUntilPredicate: removeUntilPredicate,
-      pageTransitionType: pageTransition,
-      pageTransitionSettings: pageTransitionSettings)
+          args: nyArgument,
+          navigationType: navigationType,
+          result: result,
+          removeUntilPredicate: removeUntilPredicate,
+          pageTransitionType: pageTransition,
+          pageTransitionSettings: pageTransitionSettings)
       .then((v) => onPop != null ? onPop(v) : (v) {});
 }