+++ lib/widgets/ny_stateful_widget.dart
@@ -7,7 +7,8 @@ abstract class NyStatefulWidget extends StatefulWidget {
   final String? routeName;
 
   String get route {
-    assert(routeName != null, '${this.runtimeType.toString()} is missing a routeName');
+    assert(routeName != null,
+        '${this.runtimeType.toString()} is missing a routeName');
     return (routeName ?? "");
   }